requirements.txt
fastapi[all]
pydantic[email]
sqlalchemy
uvicorn
redis
bcrypt
python-dotenv
alembic
aiohttp
python-jose
starlette
asyncpg
psycopg2-binary
celery[redis]
flower
cryptography
pyotp
qrcode[pil]
sentry-sdk[fastapi]==1.39.1
python-json-logger==2.0.7